backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI app with Chainlit integration"
requires-python = ">=3.13"
dependencies = [
    "chainlit==2.4.400",
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "nest-asyncio==1.6.0",
    "openai-agents==0.0.9",
    "pydantic==2.11.3",
    "python-dotenv==1.1.0",
    "tavily-python==0.5.4",
    "uvicorn==0.34.0",
]

[project.scripts]
open-sdk = "backend:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]